GLSL: get GLSL limping in `render-test`
The test case that is there right now is nominally a cross-compilation test, but for right now it uses the preprocessor to present completely different code for HLSL and GLSL compilation.

This change is really just fleshing out the OpenGL side of `render-test` enough that it can produce images using OpenGL to enable further testing.