Revise the (only) rewriter-error test case
This was testing that we let the downstream compiler report a parse error, but that is no longer a goal (because we need to rewrite function bodies to deal with type validation).

I've switched it to a type error in the function body, and confirmed that we do let fxc report the problem.