Ensure that lowered globals for `inout` shader parameters have unique names
If the user had a shader entry point with an `inout` parameter, we would end up lowering it to two GLSL global variables with the same name.
This change adds a `SLANG_in_` or `SLANG_out_` prefix to the two declarations.

Note: I haven't dealt with the issue that we end up printing two different `layout` qualifiers on such a variable...