Add support for dumping intermediates for debugging.
Calling:

    spSetDumpIntermedites(compileRequest, true);

will set up a mode where Slang tries to dump every intermediate HLSL, GLSL, DXBC, SPIR-V, etc. file it generates. If SPIR-V or DXBC is requested then we also dump assembly of those.

Right now the files are all named as `slang-<counter>.<ext>`, and get dropped in whatever the working directory is, but I'm open to ideas on how to improve that.

Note: this change introduces a new binary interface to `glslang`, so pulling it requires an updated `glslang.dll`.