Don't use "auto locations" mode in glslang
Fixes #88

The code was using `glslang::TShader::setAutoMapLocations` as a workaround for an old glslang issue, but apparently this mode has a bug where it ends up applying a `location` layout to the implicitly-created `gl_PerVertex` definition (which shouldn't be allowed).

This change drops the call to `setAutoMapLocations` (and `setAutoMapBindings`) since it should no longer be required.