Handle arrays when scalarizing "resources in structs"
The basic idea is that an array of `struct`s will get scalarized into per-field arrays (for any fields that need to be scalarized). So given:

    struct Foo { float x; Texture2D t; };
    cbuffer C { Foo foo[4]; }

We'll get output like:

    struct Foo { float x; };
    cbuffer C { Foo foo[4]; }
    Texture2D C_foo_t[4];

(Of course the output would also be translated over to GLSL, but I'm only concerned about this one transformation here).