Fixes for how parameter block names are set up.
We generate implicit names for global-scope parameter blocks (including HLSL `cbuffer`s, since the "name" the user sees is really just for reflection purposes), but this had a few problems:

- We used the generated names for parameter-binding purposes
  - Except for a GLSL block with an explicit name, in which case we'd use the internal name and not the reflection name for matching

- The generated named didn't match between GLSL and HLSL/Slang declarations

This change tries to fix both of these issues. I changed the name generation to try to make it identical between HLSL and GLSL (to the extent we can control it), just in case. But then I also went and changed the parameter-binding-generation logic to use the *reflection* name instead of the internal name when deciding if things are the "same" parameter.