Don't add `flat` qualifier to integer fragment output
Fixes #133

We already had logic to skip adding `flat` to a vertex input, and this just extends it to not adding `flat` to a fragment output.

Note that explicit qualifiers in the input HLSL/Slang will still be carried through to the output, so it is still possible for a Slang user to shoot themself in the foot with interpolation qualifiers.